api/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.0.4; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
markdown==3.6; python_version >= '3.8'
mysqlclient==2.2.4; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
